requirements.txt
asgiref==3.8.1
boto3==1.34.140
botocore==1.34.140
dj-database-url==2.2.0
Django==5.0.6
docopt==0.6.2
docx-mailmerge==0.5.0
jmespath==1.0.1
lxml==5.2.2
num2words==0.5.13
python-dateutil==2.9.0.post0
python-decouple==3.8
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
urllib3==2.2.2
whitenoise==6.7.0
